Refactor puppetserver container tests
This updates the puppetserver tests to use `docker-compose` instead of
`docker-run`. This also updates the tests to use the shared testing
gem from github.com/puppetlabs/pupperware.

This also includes a move from the puppet-agent-alpine to
puppet-agent-ubuntu for testing. We were seeing a lot of intermittent
network failures with the alpine container on windows (LCOW). See
https://github.com/docker/libnetwork/issues/2371 and
https://github.com/Microsoft/opengcs/issues/303 have more information on
this issue. This should hopefully clear up the intermittent name
resolution failures we were seeing.