Use PUPPETSERVER_HOSTNAME as certname if set (#2470)
Previously when logging the container's configuration, the certname was retrieved from `${SSLDIR}/certs`. This could  lead to a `file not found` issue when printing the configuration if that directory contained more than one certificate. Since the `PUPPETSERVER_HOSTNAME` environment variable is used to configure the certname for puppetserver anyway, this commit updates the logging to use that if it is set, and only fall back to searching the dir if it's not configured.